+++ app/src/test/java/org/hyperledger/besu/cli/CommandTestAbstract.java
@@ -275,6 +275,9 @@ public abstract class CommandTestAbstract {
 
   @BeforeEach
   public void initMocks() throws Exception {
+    lenient()
+        .when(mockControllerBuilderFactory.checkpoint(any()))
+        .thenReturn(mockControllerBuilderFactory);
     when(mockControllerBuilderFactory.fromEthNetworkConfig(any(), any()))
         .thenReturn(mockControllerBuilder);
     when(mockControllerBuilder.build()).thenReturn(mockController);